move ai api request to ai folder
