+++ src/handleLabeled.test.ts
@@ -2,7 +2,7 @@ import type { EmitterWebhookEvent } from "@octokit/webhooks";
 import type { Octokit } from "octokit";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { handleLabeled } from "./handleLabeled.js";
-import { runAiReview } from "./networks/ai_api_request.js";
+import { runAiReview } from "./networks/ai/ai_api_request.js";
 import { getPRFiles } from "./networks/github.js";
 import { postInlineComments } from "./networks/postInlineComment.js";
 import { postPRComment } from "./networks/postPrComment.js";
